WIP: DeterministicKey deprecate getPath(), provide path() and partialPath()
This is IMO the best way to do it. Most (but not all) existing usages of
getPath() really want partialPath(). So this commit refactors/renames
getPath() to partialPath() and then adds back the deprecated getPath()
method. fullPath() was refactored/renamed to path().

Internally, in most places we want partialPath() and this commit
reflects that. There are few cases that we might want to change.

Most external users will probably want path() (full path) but should get
a deprecation warning and can choose which one the want.

fullPath() is probably most often used in tests, see ChildKeyDerivationTest,
for example.

However, I propose merging PR #3747 as is and apply the refactoring
in this commit after we have merged the current stack of HDPath-related
changes.

Please, please don't sentence me to rebasing hell!